agg otros CU

+++ admin-starter/src/main/java/RegistrarTramite/dtos/DTOTramiteElegido.java
@@ -13,7 +13,7 @@ public class DTOTramiteElegido {
     private int codTipoTramite;
     private String nombreTipoTramite;
     private String nombreEstado;
-    private int precioTramite;
+    private double precioTramite;
     private int dniCliente;
     private String nombreCliente;
     private String apellidoCliente;
@@ -78,11 +78,11 @@ public class DTOTramiteElegido {
         this.nombreEstado = nombreEstado;
     }
 
-    public int getPrecioTramite() {
+    public double getPrecioTramite() {
         return precioTramite;
     }
 
-    public void setPrecioTramite(int precioTramite) {
+    public void setPrecioTramite(double precioTramite) {
         this.precioTramite = precioTramite;
     }
